requirements.txt
mysql_connector_repackaged==0.3.1
numpy==1.24.2
pandas==1.5.3
Pillow==9.5.0
plotly==5.13.0
plotly_express==0.4.1
requests==2.28.2
streamlit==1.20.0
streamlit_lottie==0.0.3
